requirements.txt
crewai
crewai_tools
python-dotenv
pandas
groq